remove: removing playerHealthStatus and using only PlayerStatus
