+++ Assets/Scenes/Level1.unity
@@ -1203,6 +1203,14 @@ PrefabInstance:
     m_TransformParent: {fileID: 0}
     m_Modifications:
     - target: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+      propertyPath: maxHealth
+      value: 6
+      objectReference: {fileID: 0}
+    - target: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+      propertyPath: currentHealth
+      value: 6
+      objectReference: {fileID: 0}
+    - target: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
       propertyPath: playerModes.Array.size
       value: 1
       objectReference: {fileID: 0}
@@ -12067,7 +12075,7 @@ PrefabInstance:
     - target: {fileID: 2581601332914600383, guid: cdf9c4e863253f1a6871a6f7881e1314, type: 3}
       propertyPath: healthStatus
       value: 
-      objectReference: {fileID: 1234211591}
+      objectReference: {fileID: 2102976483}
     - target: {fileID: 8558195113708807365, guid: cdf9c4e863253f1a6871a6f7881e1314, type: 3}
       propertyPath: m_Name
       value: HeartsContainerPrefab
@@ -12124,17 +12132,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   enemySpawnPoints: []
   enemyPrefabs: []
---- !u!114 &1234211591 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7679395357654442115, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
-  m_PrefabInstance: {fileID: 346729847}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 75e4775abe4417dba85ad5403bc2ebc7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1252621813
 GameObject:
   m_ObjectHideFlags: 0
@@ -30026,6 +30023,17 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 8391630477536354153, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
   m_PrefabInstance: {fileID: 346729847}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &2102976483 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+  m_PrefabInstance: {fileID: 346729847}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ef0863be3f918582b90f577d24107617, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &2136849864
 GameObject:
   m_ObjectHideFlags: 0
