+++ Assets/Scenes/Level2.unity
@@ -1202,6 +1202,14 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+      propertyPath: maxHealth
+      value: 6
+      objectReference: {fileID: 0}
+    - target: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+      propertyPath: currentHealth
+      value: 6
+      objectReference: {fileID: 0}
     - target: {fileID: 1667360318438923597, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
       propertyPath: m_Size.y
       value: 0.4217827
@@ -1213,10 +1221,6 @@ PrefabInstance:
     - target: {fileID: 6219646412613573617, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
       propertyPath: m_Name
       value: PlayerPrefab
-      objectReference: {fileID: 0}
-    - target: {fileID: 7679395357654442115, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
-      propertyPath: currentHealth
-      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 8391630477536354153, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
       propertyPath: m_LocalPosition.x
@@ -1263,6 +1267,17 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+--- !u!114 &346729848 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+  m_PrefabInstance: {fileID: 346729847}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ef0863be3f918582b90f577d24107617, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!4 &496027037 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 19105637333545446, guid: b6882751d6f9bbac982857e497205a9c, type: 3}
@@ -12149,7 +12164,7 @@ PrefabInstance:
     - target: {fileID: 2581601332914600383, guid: cdf9c4e863253f1a6871a6f7881e1314, type: 3}
       propertyPath: healthStatus
       value: 
-      objectReference: {fileID: 1234211591}
+      objectReference: {fileID: 346729848}
     - target: {fileID: 8558195113708807365, guid: cdf9c4e863253f1a6871a6f7881e1314, type: 3}
       propertyPath: m_Name
       value: HeartsContainerPrefab
@@ -12206,17 +12221,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   enemySpawnPoints: []
   enemyPrefabs: []
---- !u!114 &1234211591 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7679395357654442115, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
-  m_PrefabInstance: {fileID: 346729847}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 75e4775abe4417dba85ad5403bc2ebc7, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1252621813
 GameObject:
   m_ObjectHideFlags: 0
