+++ Assets/Scripts/Enemies/Orcs/IOrcAttack.cs
@@ -5,5 +5,5 @@ public interface IOrcAttack
     void ApplyDamage();
     void HandleTriggerEnter(OrcScript context, Collider2D collision);
 
-    void SetLastTarget(PlayerHealthStatus target);
+    void SetLastTarget(PlayerStatus target);
 }
