--- Assets/Scripts/Player/Status/PlayerHealthStatus.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 75e4775abe4417dba85ad5403bc2ebc7
